Integrate EKF test functions into main.cpp and fix CMake issues
- Remove test_ekf_simple.cpp file and integrate EKF validation tests into main.cpp
- Add configurable RUN_EKF_TESTS flag to control test execution
- Add missing encoderMeasurementModel() and encoderMeasurementJacobian() functions to SensorFusion.cpp
- Update CMakeLists.txt to remove test_ekf_simple target and fix build dependencies
- Fix CMake build issues and eliminate duplicate main function conflicts

The EKF tests can now be run from the main executable by setting RUN_EKF_TESTS=true
